+++ bin/commands/install.js
@@ -1,14 +1,14 @@
 module.exports = {
   registerVorpalCommand: (vorpal, configuration) => {
     return vorpal
-      .command(`audit`)
-      .alias(`a`)
-      .description(`Audit internal rules.`)
+      .command(`install`)
+      .alias(`i`)
+      .description(`Install a package.`)
       .validate(function (args) {
         return true
       })
       .action((args, done) => {
-        vorpal.log(`pending...`)
+
         done()
       })
   }
